kino/riff.cc: suppress compiler warning
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@1373 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/modules/kino/riff.cc
@@ -518,7 +518,7 @@ void RIFFFile::ParseChunk( int parent )
 	read( fd, &type, sizeof( type ) );
 	if ( type == make_fourcc( "LIST" ) )
 	{
-		typesize = -sizeof( type );
+		typesize = (int) -sizeof( type );
 		fail_if( lseek( fd, typesize, SEEK_CUR ) == ( off_t ) - 1 );
 		ParseList( parent );
 	}
